+++ dwds/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Update to be forward compatible with changes to `package:shelf_web_socket`.
 - Added support for some debugging APIs with the DDC library bundle format. - [#2537](https://github.com/dart-lang/webdev/issues/2537),[#2544](https://github.com/dart-lang/webdev/issues/2544)
 
+
 ## 24.2.0
 
 - Consolidate `FrontendServerDdcStrategyProvider` and `FrontendServerRequireStrategyProvider` under a shared parent class. - [#2517](https://github.com/dart-lang/webdev/issues/2517)