+++ dwds/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Update to be forward compatible with changes to `package:shelf_web_socket`.
 - Added support for some debugging APIs with the DDC library bundle format. - [#2537](https://github.com/dart-lang/webdev/issues/2537)
+- Added support for some debugging APIs with the DDC library bundle format. - [#2537](https://github.com/dart-lang/webdev/issues/2537),[#2544](https://github.com/dart-lang/webdev/issues/2544)
 - Expose a partial implementation of
   `FrontendServerDdcLibraryBundleStrategyProvider`.
 - Update `package:vm_service_interface` to '^2.0.1'.