+++ dwds/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 24.2.1-wip
 
 - Update to be forward compatible with changes to `package:shelf_web_socket`.
-- Added support for some debugging APIs with the DDC library bundle format. - [#2537](https://github.com/dart-lang/webdev/issues/2537),[#2544](https://github.com/dart-lang/webdev/issues/2544)
+- Added support for some debugging APIs with the DDC library bundle format. - [#2537](https://github.com/dart-lang/webdev/issues/2537),[#2544](https://github.com/dart-lang/webdev/issues/2544),[#2548](https://github.com/dart-lang/webdev/issues/2548)
 
 ## 24.2.0